Fix for the 3.8 version + some adjustements
Hey, I did a little mistake for the 3.8 and the adaptation for the
version 1.11 of Minecraft, the plugin did not work anymore.

I did the best of I can for publish to you a fix and there he is :)
I did by the way many adjustements like :
The invicible time when you do a /srtp the time before the update (for
the effects) was 100 milliseconds, and now it's 50 milliseconds.
The radius of the random teleportation was 1000 blocks and now 10000
blocks
I am sorry for the caused...